foodtrace/products/requirements.txt
flask
web3
qrcode
python-dotenv
hexbytes
py-solc-x